requirements.txt
runpod
torch
transformers
accelerate
pillow
numpy
av
